Move BayesianScorer into belief and fix bugs
